bin/webserver/app/requirements.txt
python-dotenv
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
pymysql